Remove struct fctype_t
It is only ever allocated by eppic_addfunc_ctype() and added to a ctypes
field of struct fdata, then it is never used, except it gets properly
cleaned up in eppic_freefile().

It looks like an unfinished idea, but since the intention is not clear, it
seems best to remove it.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>