Improve parsing of integer constants
Some integer constants are not parsed according to the ISO C standard:

- Decimal constants are always parsed as signed. If the actual constant
  overflows a (signed) long long, the conversion is incorrect.

- Invalid suffixes are not properly rejected (e.g. multiple 'u' suffixes).

- Invalid digits between the last valid digit and the suffix are silently
  ignored, e.g. "0129U" is accepted as 012U.

Let's do the parsing in logical steps.

First, determine if the number is decimal, octal, or hexadecimal, and skip
the prefix if necessary. Note that the leading zero is not skipped for
octal constants. This allows to parse "0" as an octal constant.

Second, parse the suffix(es) to find out the type.

Third, parse the digit sequence. Note that all constants are in fact
positive. If a minus sign is present, it is parsed as a unary operator
outside the lexer.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>