Fix a user-after-free issue in eppic_deletefile()
There is a use-after-free issue noticed by libasan in
eppic_deletefile(). The root cause is, eppic_freefile(fd) will free
fd->fname and fd itself, and eppic_findfile(name, 1) will access fd->name
and fd->next. When name and fd are associated with the same file, this
is a use-after-free issue which will trigger eppic crash.

This patch will reorder the eppic_freefile and eppic_findfile within
eppic_deletefile, so the bug will not retrigger.

Signed-off-by: Tao Liu <ltao@redhat.com>